requirements.txt
bidict==0.21.2
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.1
Flask-SocketIO==5.0.3
gunicorn==20.1.0
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycryptodome==3.10.1
pymongo==3.11.4
python-engineio==4.2.0
python-socketio==5.3.0
scrypt==0.8.18
simple-websocket==0.2.0
Werkzeug==2.0.1
wsproto==1.0.0
